requirements.txt
streamlit==1.35.0
selenium==4.20.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.2
google-api-core>=2.15.0